refactor(release-cascade): move and rename API and form handler
